+++ dbms/src/Flash/tests/gtest_spill_join.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <Flash/tests/gtests_join.h>
+#include <Flash/tests/gtest_join.h>
 
 namespace DB
 {
